requirements.txt
pandas
requests
openpyxl
ipykernel
fastapi
uvicorn
pydantic
anyio
python-jose
aiofiles
pydantic-settings
asgi-correlation-id
python-multipart
python-dotenv
aiohttp
watchtower
oracledb